requirements.txt
discord.py==2.3.2
google-generativeai==0.3.2
selenium==4.15.2
Pillow==10.1.0
python-dotenv==1.0.0
beautifulsoup4==4.12.2
requests==2.31.0
webdriver-manager==4.0.1
undetected-chromedriver==3.5.4